[OPT] Implement general eigenvalue decomposition in PyTorch (#42)
To invert the sum of two Kronecker products, one needs to solve a generalized eigenvalue problem.
PyTorch does not offer an API to do so, which is why currently we are loading all matrices to CPU and use `scipy.linalg.eigh` to solve the generalized eigenproblem there. 
This costs a GPU-CPU sync when we are running on a CUDA device.
This PR adds a custom implementation of `eigh` that runs fully in PyTorch.
However, I am not changing the default method to this one yet, as I did not observe a performance difference on the problems we are currently running (probably my custom implementation is still naive compared to `scipy`'s).
The `scipy` implementation seems to be slightly more stable (I could use a smaller damping value of `1e-12` on Poisson2d while I had to use `1e-10` with the `torch` implementation).
For big models, we need to re-assess if it is worth to change to the `torch` implementation to avoid GPU-CPU syncs. 

* [ADD] Implement general eigenvalue decomposition in PyTorch

* [ADD] Change default backend to torch

* [ADD] Debugging script to check speed-up [no ci]

* [ADD] Change interface of `eigh` to return inverse of B, too

* [DEL] Debugging script